chore(Datagrid): clean up column logic into new component
