+++ packages/cloud-cognitive/src/components/Datagrid/Datagrid/addons/CustomizeColumns/Columns.js
@@ -12,9 +12,8 @@ import { HTML5Backend } from 'react-dnd-html5-backend';
 import { Checkbox } from 'carbon-components-react';
 import update from 'immutability-helper';
 import { pkg } from '../../../../../settings';
-import DraggableElement from '../../DraggableElement';
-import { isColumnVisible } from './common';
 import classNames from 'classnames';
+import { DraggableItemsList } from './DraggableItemsList';
 
 const blockClass = `${pkg.prefix}--datagrid`;
 
@@ -116,127 +115,18 @@ const Columns = ({
               labelText={selectAllLabel}
             />
           </div>
-          {columns
-            // hide the columns without Header, e.g the sticky actions, spacer
-            .filter((colDef) => {
-              const sortableTitle =
-                isTableSortable && colDef.Header().props.children.props?.title;
-              return (
-                (!!colDef.Header.props && !!colDef.Header.props?.title) ||
-                (isTableSortable && !!sortableTitle)
-              );
-            })
-            .filter((colDef) => !colDef.isAction)
-            .filter((colDef) => {
-              const sortableTitle =
-                isTableSortable && colDef.Header().props.children.props?.title;
-              return (
-                filterString.length === 0 ||
-                ((isTableSortable
-                  ? sortableTitle?.toLowerCase().includes(filterString)
-                  : colDef.Header.props?.title
-                      ?.toLowerCase()
-                      .includes(filterString)) &&
-                  colDef.id !== 'spacer')
-              );
-            })
-            .map((colDef, i) => {
-              const isSortableColumn = !!colDef.canSort && !!isTableSortable;
-              const sortableTitle =
-                isTableSortable && colDef.Header().props.children.props?.title;
-              const searchString = new RegExp('(' + filterString + ')');
-              const res = filterString.length
-                ? isSortableColumn
-                  ? sortableTitle.toLowerCase().split(searchString)
-                  : colDef.Header.props.title.toLowerCase().split(searchString)
-                : null;
-              const firstWord =
-                res !== null
-                  ? res[0] === ''
-                    ? res[1].charAt(0).toUpperCase() + res[1].substring(1)
-                    : res[0].charAt(0).toUpperCase() + res[0].substring(1)
-                  : null;
-              const highlightedText =
-                res !== null
-                  ? res[0] === ''
-                    ? `<strong>${firstWord}</strong>` + res[2]
-                    : firstWord + `<strong>${res[1]}</strong>` + res[2]
-                  : isSortableColumn
-                  ? sortableTitle
-                  : colDef.Header.props?.title;
-              const isFrozenColumn = !!colDef.sticky;
-              const listContents = (
-                <>
-                  <Checkbox
-                    wrapperClassName={`${blockClass}__customize-columns-checkbox-wrapper`}
-                    checked={isColumnVisible(colDef)}
-                    disabled={isFrozenColumn}
-                    onChange={onSelectColumn.bind(null, colDef)}
-                    id={`${blockClass}__customization-column-${colDef.id}`}
-                    labelText={
-                      isSortableColumn
-                        ? sortableTitle
-                        : colDef.Header.props?.title
-                    }
-                    title={
-                      isSortableColumn
-                        ? sortableTitle
-                        : colDef.Header.props?.title
-                    }
-                    className={`${blockClass}__customize-columns-checkbox`}
-                    hideLabel
-                  />
-                  {
-                    <div
-                      dangerouslySetInnerHTML={{ __html: highlightedText }}
-                    ></div>
-                  }
-                </>
-              );
-
-              return (
-                <DraggableElement
-                  key={colDef.id}
-                  index={i}
-                  listData={columns}
-                  setListData={setColumnsObject}
-                  id={`dnd-datagrid-columns-${colDef.id}`}
-                  type="column-customization"
-                  disabled={filterString.length > 0 || isFrozenColumn}
-                  ariaLabel={
-                    isSortableColumn
-                      ? sortableTitle
-                      : colDef.Header.props?.title
-                  }
-                  onGrab={setAriaRegionText}
-                  isFocused={focusIndex === i}
-                  moveElement={moveElement}
-                  onArrowKeyDown={(e, isGrabbed, currentIndex) => {
-                    if (isGrabbed) {
-                      const nextIndex = getNextIndex(
-                        columns,
-                        currentIndex,
-                        e.key
-                      );
-                      e.preventDefault();
-                      e.stopPropagation();
-
-                      if (nextIndex >= 0 && !columns[nextIndex]?.sticky) {
-                        setFocusIndex(nextIndex);
-                        moveElement(currentIndex, nextIndex);
-                        e.target.scrollIntoView({
-                          block: 'center',
-                        });
-                      }
-                    }
-                  }}
-                  isSticky={isFrozenColumn}
-                  selected={isColumnVisible(colDef)}
-                >
-                  {listContents}
-                </DraggableElement>
-              );
-            })}
+          <DraggableItemsList
+            columns={columns}
+            filterString={filterString}
+            focusIndex={focusIndex}
+            getNextIndex={getNextIndex}
+            isTableSortable={isTableSortable}
+            moveElement={moveElement}
+            onSelectColumn={onSelectColumn}
+            setAriaRegionText={setAriaRegionText}
+            setColumnsObject={setColumnsObject}
+            setFocusIndex={setFocusIndex}
+          />
         </ol>
       </DndProvider>
     </div>
